feat: taking into account PR review, create utils directory, add flatlist, separate CSS, close modal
