+++ components/AddActivity/AddActivityButton.tsx
@@ -1,0 +1,17 @@
+import { View, Text, TouchableOpacity } from 'react-native';
+import {styles} from "./AddActivityButton.style.ts";
+
+type AddActivityButtonProps = {
+    onPress: () => void;
+};
+export const AddActivityButton = ({ onPress }:AddActivityButtonProps) => {
+
+  return (
+          <View style={styles.buttonContainer}>
+              <TouchableOpacity style={styles.button} onPress={onPress} activeOpacity={0.7}>
+                  <Text style={styles.buttonText}>+</Text>
+              </TouchableOpacity>
+          </View>
+  );
+
+};
