+++ components/WorkInput/WorkInput.style.ts
@@ -1,0 +1,33 @@
+import {StyleSheet} from "react-native";
+
+export const styles = StyleSheet.create({
+    inputContainer: {
+        marginBottom: 16,
+        backgroundColor: '#fff',
+        padding: 8,
+        borderRadius: 4,
+    },
+    inputTitle: {
+        fontSize: 16,
+        fontWeight: 'bold',
+        marginBottom: 4,
+    },
+    textInput: {
+        height: 40,
+        borderColor: '#ccc',
+        borderWidth: 1,
+        borderRadius: 4,
+        paddingHorizontal: 8,
+    },
+    deleteButton: {
+        marginTop: 8,
+        padding: 8,
+        backgroundColor: '#ff4d4d',
+        borderRadius: 4,
+        alignItems: 'center',
+    },
+    deleteButtonText: {
+        color: '#fff',
+        fontWeight: 'bold',
+    },
+});
